+++ src/current/mod.rs
@@ -7,5 +7,5 @@ pub(super) use constant_current::ConstantCurrent;
 
 pub(crate) trait CurrentData {
     /// Return the current (u, v) at the given (x, y)
-    fn current(&self, x: &f32, y: &f32) -> Result<(f32, f32), Error>;
+    fn current(&self, x: &f64, y: &f64) -> Result<(f64, f64), Error>;
 }
